backend/requirements.txt
Django==3.2.23
django-cors-headers==3.13.0
djangorestframework==3.12.4
pymysql==1.1.1
python-dotenv==1.1.1
mysqlclient==2.2.7
pytz==2025.2
sqlparse==0.5.3
asgiref==3.10.0